Refactor TreeImpostorAtlas into smaller files
Split the large TreeImpostorAtlas class into smaller, more focused files:
- ImpostorTypes.h: Data structs (OctahedralAtlasConfig, TreeImpostorArchetype, TreeLODSettings)
- OctahedralMapping.h: Hemi-octahedral encode/decode utility functions
- TreeImpostorAtlasCapture.cpp: Pipeline creation and cell rendering code
- TreeImpostorAtlas.cpp: Core atlas management and archetype generation (slimmed down)
